Removed about and added content to index
